+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Changed:
 Fixed:
 ^^^^^^
 * vcf2cytosure container https://github.com/Clinical-Genomics/BALSAMIC/pull/1159
+* Link external fastqs to case folder & create case directory https://github.com/Clinical-Genomics/BALSAMIC/pull/1195
 
 Removed:
 ^^^^^^^